make find_ancestors and find_descendants less stateful
This makes the find_ancestors and find_descendants less stateful by
passing around the list of person handles they find. This makes it so
that if the two passes over the data encounter the same person (namely
the active person) they won't conflict and cause it to bail too early.