GetGOV addon update
I made another update that adds an Alternate Name in the preferred language if there is exactly one Name for the place (no Alternate Names). This makes the Title work most of the time. But is not technically correct (for example, Württemberg is not english, but gets added as an Alternate Name with the language set to 'en' if your preferred language is 'en').

https://gramps-project.org/bugs/view.php?id=8974#c49822