Fix issue with Notes in FamilyTree graphical report (#158)
Notes with multiple lines attached to  People records can cause the Family Tree graphical report to output the notes outside the boundary of the drawn box or overwrite other text within the drawn box.  To resolve this any carriage returns and new lines are replaced with spaces prior to outputing in the report.