Minor enhancements and fixes for GenealogyTree[gramps50] (#184)
* GenealogyTree: Correctly handle max_up

If there is more data than the user wants to export (specified by
the 'max_up' parameter), correctly emit a termination node instead
of omitting the last level. The check is analogous to the case in
subgraph_down.

* GenealogyTree: Add option to exclude siblings of ancestors

This reduces some space problems for sandclock reports.

* GenealogyTree: Allow no generations up or down

This makes sense for Sandclock reports to end immediately above /
below the center person / family.