Add new combined view
An experimental view based on the relationship view.  It displays
both relationships and events for a given person.  More detail is
displayed including attributes and quick access to source media.

This is still a work in progress, but is being relaesed for user
feedback.  In particular different navigation needs tobe written.