Fixes #11596 GetGOV language setting is case sensitive
Gramps Preferences > Display > Place Format's language setting allows input in mixed case, but GetGOV addon interprets it correctly only when the language code is in lower case otherwise it defaults to German ('de').

This change converts the value of the language field to lower case when used by the addon so that GetGOV accepts it and returns results in that language (as far as possible).