Improve the robustness of finding form events
Previously, for a given form event, the citation was stored.
If later, the form was edited, an attempt was made to locate the correct event, given just the citation.
If the citation was shared by multiple events, this was prone to return a different event than we started with.

This change tracks the original form event and passes it into EditForm directly, removing the need for the reverse lookup.