Todo report context (#598)
* Add all dates to person output

The dates are grouped by type, each date has it's own row.
This can help give context to the note.

* Add alternate names to todo report

* Move the dates to the right 1 column

This works nicely as an indent so that it's clear this information
belongs with the person.

* Sort events based on type and then date

* Output all dates for families

Refactored code for writing out all events so that it can be used for
both Person and Family.

* Enhance output of events

* Add all people referenced by the event
* Add all families referenced by the event

* Add citation

* Add media

* Add source and repository

* Add sort support for the new supported object types

* Use database constants instead of strings for class names

This reduces the changes to have a typo in the code.

* Add option to specify the title

* Add the ability to filter by note type

This now allows one to specify all notes of type "To Do" in the entire
database. One can now also combine the tag and note type filters in
the options dialog.