Fix edge case on common ancestor calc and added visual phasing (#605)
* Fix edge case on common ancestor calc

Fix the edge case where the relationship between 2 people can be determined but there is no common ancestor in the tree. Specifically the relationship leverages siblings where the parent does not exist in the tree.

* Add grandparent view

Add a view that is the grandparent phasing. Each segment that has a hared grandparent is painted the same. There are 4 colors use: paternal grandmother, paternal grandfather, maternal grandmother, and maternal grandfather. 

Segments that do not have a known grandparent are not displayed.

* Add great grand visual phasing

Add but disable the following capabilities

* user-config for Association Type

* Additional tweaks for TSV output and visual phasing

Add config for TSV output
Reworked visual phasing code for performance