Form references[v52] (#265)
* Add the concept of a default reference to the Form gramplet.

The intended use it for the default reference in the form definition to provide a template, which the user manually completes.
For example for the UK1841 form, the default reference could be
'Department: HO, Series: 107, Piece: , Book: , Folio: , Page: , Line: '
which the user then manually completes to create
'Department: HO, Series: 107, Piece: 123, Book: 15, Folio: 12, Page: 3, Line: 10-12'

* When displaying a new form, use the default reference text specified in the form defintion.

* Add default reference templates for English and Welsh censuses.
Reference templates conform to the WHERE_WITHIN_SOURCE format of the GEDCOM 5.5.1 specification

* Add default references for English Birth, Marriage and Death certificates.

Note 1: In some cases the default reference fields are also included as heading fields. This is deliberate. If the existing heading fields are removed, any existing data in these fields will no longer display.

* Fix typo