Polish, cleanup around return, LOG and UI (unfinished)
Now, it "prints" JSONL content via the TextBuffer on gramplet. So, we can also read and copy from gramps UI. The ERROR dialogs, inherited from lxml (python), are not "children dialogs" into gramps UI manager. So, they can pop-up if the parsing failed. A proper solution should exist, but as they will raise an error anyway, we only get them during debug or when need a fix (either on code or on user side).